[BE] feat: domain Menu, MenuType, SaleType 설계

+++ BE/src/main/java/sidedish/jbc/dto/ResponseMenu.java
@@ -1,18 +1,16 @@
 package sidedish.jbc.dto;
 
-import java.math.BigDecimal;
-
 public class ResponseMenu {
 
 	private long id;
 	private String name;
 	private String description;
-	private BigDecimal price;
+	private int price;
 	private int menuType;
 	private int saleType;
 	private String imageURL;
 
-	public ResponseMenu(long id, String name, String description, BigDecimal price, int menuType,
+	public ResponseMenu(long id, String name, String description, int price, int menuType,
 		int saleType, String imageURL) {
 		this.id = id;
 		this.name = name;
